Organize tests by responsibility and move architecture checks to lint
